+++ .env
@@ -1,18 +1,17 @@
 Setting the path for the FFMPEG.
 FFMPEG ="ffmpeg"
-Setting the path for the ImageMagick.
+Setting the path for ImageMagick.
 MAGICK="magick"
 
-
 Setting the path for the input files.
-UPLOADS_PATH_PDF="./src/converters/pdfConverter/inputs/"
-UPLOADS_PATH_AUDIO="./src/converters/audioConverter/inputs"
-UPLOADS_PATH_IMAGE="./src/converters/imageConverter/inputs/"
-UPLOADS_PATH_VIDEO="./src/converters/videoConverter/inputs/"
-
+UPLOADS_PATH_PDF="./src/service/pdfConverter/inputs/"
+UPLOADS_PATH_AUDIO="./src/service/audioConverter/inputs/"
+UPLOADS_PATH_IMAGE="./src/service/imageConverter/inputs/"
+UPLOADS_PATH_VIDEO="./src/service/videoConverter/inputs/"
 
 Setting the path for the output files.
-DOWNLOAD_PATH_AUDIO="./src/converters/audioConverter/outputs/"
-DOWNLOAD_PATH_IMAGE="./src/converters/imageConverter/outputs/"
-DOWNLOAD_PATH_VIDEO="./src/converters/videoConverter/outputs/"
-DOWNLOAD_PATH_PDF="./src/converters/pdfConverter/outputs/"
+DOWNLOAD_PATH_AUDIO="./src/service/audioConverter/outputs/"
+DOWNLOAD_PATH_IMAGE="./src/service/imageConverter/outputs/"
+DOWNLOAD_PATH_VIDEO="./src/service/videoConverter/outputs/"
+DOWNLOAD_PATH_PDF="./src/service/pdfConverter/outputs/"
+
