+++ README.md
@@ -1,22 +1,56 @@
 # AT20_COMPILER_SERVICE
-DEPENDENCIES FOR CONVERTERS
-download https://www.gyan.dev/ffmpeg/builds/ffmpeg-git-full.7z
-install and modify env variables
-install imageMagick
-install ghostscript
-download https://ghostscript.com/releases/index.html
-download the first link here documentation
-https://ghostscript.com/docs/9.54.0/Install.htm
-download the General Public Release for your PC
-ubuntu
-rpm -Uvh ImageMagick-7.1.1-3.x86_64.rpm
-rpm -Uvh ImageMagick-libs-7.1.1-3.x86_64.rpm
-you can look to the page for more information
-https://imagemagick.org/script/download.php
+
+## Available Scripts
+
+In the project directory, you can run:
+
 npm install or npm update
-download file
-$npm i --save lodash
-$npm install body-parser
-$ npm install express
-$ npm install --save multer
-$ npm install express-fileupload
+
+# download file
+
+npm i --save lodash
+
+npm install body-parser
+
+npm install express
+
+npm install --save multer
+
+npm install express-fileupload
+
+In the project directory, you can run:
+
+### `npm server.js`
+
+Runs the app in the development mode.\
+Open [http://localhost:9090](http://localhost:9090) to view it in your browser.
+
+
+## DEPENDENCIES FOR CONVERTERS
+
+download https://www.gyan.dev/ffmpeg/builds/ffmpeg-git-full.7z
+
+ install and modify env variables
+
+ install imageMagick
+
+ install ghostscript
+
+download https://ghostscript.com/releases/index.html
+
+# download the first link here documentation
+
+https://ghostscript.com/docs/9.54.0/Install.htm
+
+download the General Public Release for your PC
+
+rpm -Uvh ImageMagick-7.1.1-3.x86_64.rpm
+
+rpm -Uvh ImageMagick-libs-7.1.1-3.x86_64.rpm
+
+# You can look to the page for more information
+
+https://imagemagick.org/script/download.php
+
+que ahce el servicio
+anotar
