+++ src/service/mocks/imageConvert.js
@@ -1,13 +1,13 @@
-/**
-@imageConvert.js Copyright(c) 2023 Jalasoft
-2643 Av Melchor Perez de Olguin, Colquiri Sud, Cochabamba, Bolivia.
-Av.General Inofuentes esquina Calle20, Edificio Union No1376, La Paz, Bolivia
-All rights reserved
-This software is the confidential and proprietary information of
-Jalasoft,ConfidentialInformation"). You shall not
-disclose such Confidential Information and shall use it only in
-accordance with the terms of the license agreement you entered into
-with Jalasoft
+/*
+* @imageConvert.js Copyright(c) 2023 Jalasoft
+* 2643 Av Melchor Perez de Olguin, Colquiri Sud, Cochabamba, Bolivia.
+* Av.General Inofuentes esquina Calle20, Edificio Union No1376, La Paz, Bolivia
+* All rights reserved
+* This software is the confidential and proprietary information of
+* Jalasoft,ConfidentialInformation"). You shall not
+* disclose such Confidential Information and shall use it only in
+* accordance with the terms of the license agreement you entered into
+* with Jalasoft
 */
 const dotenv = require('dotenv');
 const { ImageCommand } = require('../imageConverter/imageCommand.js');
@@ -16,6 +16,7 @@ const { Execute } = require('../Execute.js');
 dotenv.config({path: __dirname + './../../../.env'});
 
 const newImageFile = 'test.jpg';
+const newImageFileName='test';
 const imageOutExtension = 'jpg';
 //Creates a new object image for image commands
 var image = new ImageCommand();
@@ -33,7 +34,7 @@ image.typeOfOutput = 'grayscale';
 //Sets the degrees to rotate clock wisw CW
 image.rotateCW = 90;
 //Creates the output path according to design
-const imageOutFilePath = `${'../imageConverter/outputs/'}${image.fileName(newImageFile)}.${imageOutExtension}`;
+const imageOutFilePath = `${'../imageConverter/outputs/'}${newImageFileName}.${imageOutExtension}`;
 image.convertedFilePath = imageOutFilePath;
 //Gets the command to execute the desired action
 var command = image.getCommand();
